+++ graphics/generators.go
@@ -5,7 +5,7 @@ import (
 )
 
 //ConvertTuple Function
-func convertTuple(x, y, z float64, w int) (interface{}, error) {
+func convertTuple(x, y, z float32, w int) (interface{}, error) {
 	if w == 1 {
 		return Point{X: x, Y: y, Z: z, W: w}, nil
 	} else if w == 0 {
@@ -16,16 +16,16 @@ func convertTuple(x, y, z float64, w int) (interface{}, error) {
 }
 
 //NewTuple Function
-func NewTuple(x, y, z float64, w int) Tuple {
+func NewTuple(x, y, z float32, w int) Tuple {
 	return Tuple{X: x, Y: y, Z: z, W: w}
 }
 
 //NewVector Function
-func NewVector(x, y, z float64) Vector {
+func NewVector(x, y, z float32) Vector {
 	return Vector{X: x, Y: y, Z: z, W: 0}
 }
 
 //NewPoint Function
-func NewPoint(x, y, z float64) Point {
+func NewPoint(x, y, z float32) Point {
 	return Point{X: x, Y: y, Z: z, W: 1}
 }
